1) Changed floats to doubles for easy use of apache clustering methods 2) Moved embedding iterators to embedding class 3) Started on tests for clustering using dbscan 4) In order to not slow down the machine during run, we use fixed thread pools only

+++ src/main/java/org/uu/nl/embedding/CoOccurrenceMatrix.java
@@ -10,7 +10,7 @@ public interface CoOccurrenceMatrix {
 	double max();
 	int cIdx_I(int i);
 	int cIdx_J(int j);
-	float cIdx_C(int i);
+	double cIdx_C(int i);
 	int coOccurrenceCount();
 	void shuffle();
 }
